requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==4.1.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.4
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.0.4
command-not-found==0.3
cryptography==3.4.8
dataclasses-json==0.6.7
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
entrypoints==0.4
exceptiongroup==1.2.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.14.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.7
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-text-splitters==0.2.2
langsmith==0.0.87
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.0.5
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
openai==1.35.4
orjson==3.10.5
packaging==23.2
pandas==2.2.2
pillow==10.3.0
pipdeptree==2.2.0
platformdirs==4.2.2
protobuf==3.20.1
pyarrow==16.1.0
pycairo @ file:///home/mhc/MHC_Chatbot/pycairo
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
Pympler==1.0.1
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==2.4.7
PyPDF2==1.26.0
PyPDF4==1.27.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.1
PyYAML==5.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
SecretStorage==3.3.1
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.0
streamlit==1.8.0
systemd-python==234
tenacity==8.4.2
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
validators==0.28.3
virtualenv==20.26.3
wadllib==1.3.6
watchdog==4.0.1
yarl==1.9.4
zipp==1.0.0
